Security

+++ src/main/java/com/example/Libreria/CrudRepository/UsuarioCrudRepository.java
@@ -2,6 +2,7 @@ package com.example.Libreria.CrudRepository;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Stream;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
@@ -13,9 +14,14 @@ import com.example.Libreria.Entidades.Usuario;
 
 
 @Repository
-public interface UsuarioCrudRepository extends CrudRepository<Usuario, Long>{
+public interface UsuarioCrudRepository extends CrudRepository<Usuario, Integer>{
 	
 	List<Usuario> findByName(@Param("name") String name);
+	
+	/* Optional<Usuario> findById(@Param("id")Integer id); */
+
+
+	
 	
 	
 }
